Revert "Started adjusting uvs to match pixel snapping. (#162049)"
This reverts commit dc580aff7885d4848960d658ed63345e4838e1c6.
