+++ engine/src/flutter/impeller/display_list/BUILD.gn
@@ -78,10 +78,6 @@ template("display_list_unittests_component") {
     "dl_playground.cc",
     "dl_playground.h",
     "dl_unittests.cc",
-    "testing/render_text_in_canvas.cc",
-    "testing/render_text_in_canvas.h",
-    "testing/rmse.cc",
-    "testing/rmse.h",
   ]
   additional_sources = []
   if (defined(invoker.sources)) {
