+++ engine/src/flutter/impeller/display_list/dl_golden_blur_unittests.cc
@@ -6,17 +6,57 @@
 
 #include "flutter/display_list/dl_builder.h"
 #include "flutter/display_list/effects/dl_mask_filter.h"
-#include "flutter/impeller/display_list/testing/render_text_in_canvas.h"
-#include "flutter/impeller/display_list/testing/rmse.h"
 #include "flutter/impeller/geometry/round_rect.h"
 #include "flutter/impeller/golden_tests/screenshot.h"
 #include "flutter/testing/testing.h"
 #include "gtest/gtest.h"
+#include "impeller/typographer/backends/skia/text_frame_skia.h"
+#include "txt/platform.h"
 
 namespace flutter {
 namespace testing {
 
 using impeller::Font;
+
+namespace {
+struct TextRenderOptions {
+  bool stroke = false;
+  DlScalar font_size = 50;
+  DlColor color = DlColor::kYellow();
+  std::shared_ptr<DlMaskFilter> mask_filter;
+};
+
+bool RenderTextInCanvasSkia(DlCanvas* canvas,
+                            const std::string& text,
+                            const std::string_view& font_fixture,
+                            DlPoint position,
+                            const TextRenderOptions& options = {}) {
+  auto c_font_fixture = std::string(font_fixture);
+  auto mapping = flutter::testing::OpenFixtureAsSkData(c_font_fixture.c_str());
+  if (!mapping) {
+    return false;
+  }
+  sk_sp<SkFontMgr> font_mgr = txt::GetDefaultFontManager();
+  SkFont sk_font(font_mgr->makeFromData(mapping), options.font_size);
+  auto blob = SkTextBlob::MakeFromString(text.c_str(), sk_font);
+  if (!blob) {
+    return false;
+  }
+
+  auto frame = impeller::MakeTextFrameFromTextBlobSkia(blob);
+
+  DlPaint text_paint;
+  text_paint.setColor(options.color);
+  text_paint.setMaskFilter(options.mask_filter);
+  // text_paint.mask_blur_descriptor = options.mask_blur_descriptor;
+  // text_paint.stroke_width = 1;
+  // text_paint.style =
+  //     options.stroke ? Paint::Style::kStroke : Paint::Style::kFill;
+  canvas->DrawTextFrame(frame, position.x, position.y, text_paint);
+  return true;
+}
+
+}  // namespace
 
 TEST_P(DlGoldenTest, TextBlurMaskFilterRespectCTM) {
   impeller::Point content_scale = GetContentScale();
@@ -71,6 +111,41 @@ TEST_P(DlGoldenTest, TextBlurMaskFilterDisrespectCTM) {
 
   ASSERT_TRUE(OpenPlaygroundHere(builder.Build()));
 }
+
+namespace {
+double CalculateDistance(const uint8_t* left, const uint8_t* right) {
+  double diff[4] = {
+      static_cast<double>(left[0]) - right[0],  //
+      static_cast<double>(left[1]) - right[1],  //
+      static_cast<double>(left[2]) - right[2],  //
+      static_cast<double>(left[3]) - right[3]   //
+  };
+  return sqrt((diff[0] * diff[0]) +  //
+              (diff[1] * diff[1]) +  //
+              (diff[2] * diff[2]) +  //
+              (diff[3] * diff[3]));
+}
+
+double RMSE(const impeller::testing::Screenshot* left,
+            const impeller::testing::Screenshot* right) {
+  FML_CHECK(left);
+  FML_CHECK(right);
+  FML_CHECK(left->GetWidth() == right->GetWidth());
+  FML_CHECK(left->GetHeight() == right->GetHeight());
+
+  int64_t samples = left->GetWidth() * left->GetHeight();
+  double tally = 0;
+
+  const uint8_t* left_ptr = left->GetBytes();
+  const uint8_t* right_ptr = right->GetBytes();
+  for (int64_t i = 0; i < samples; ++i, left_ptr += 4, right_ptr += 4) {
+    double distance = CalculateDistance(left_ptr, right_ptr);
+    tally += distance * distance;
+  }
+
+  return sqrt(tally / static_cast<double>(samples));
+}
+}  // namespace
 
 // This is a test to make sure that we don't regress "shimmering" in the
 // gaussian blur. Shimmering is abrupt changes in signal when making tiny
