+++ engine/src/flutter/impeller/display_list/dl_golden_unittests.cc
@@ -8,8 +8,6 @@
 #include "display_list/dl_paint.h"
 #include "display_list/geometry/dl_geometry_types.h"
 #include "flutter/display_list/dl_builder.h"
-#include "flutter/impeller/display_list/testing/render_text_in_canvas.h"
-#include "flutter/impeller/display_list/testing/rmse.h"
 #include "flutter/impeller/geometry/path_builder.h"
 #include "flutter/testing/testing.h"
 #include "gtest/gtest.h"
@@ -350,68 +348,6 @@ TEST_P(DlGoldenTest, SaveLayerAtFractionalValue) {
   builder.Restore();
 
   ASSERT_TRUE(OpenPlaygroundHere(builder.Build()));
-}
-
-namespace {
-int32_t CalculateMaxY(const impeller::testing::Screenshot* img) {
-  const uint32_t* ptr = reinterpret_cast<const uint32_t*>(img->GetBytes());
-  int32_t max_y = 0;
-  for (uint32_t i = 0; i < img->GetHeight(); ++i) {
-    for (uint32_t j = 0; j < img->GetWidth(); ++j) {
-      uint32_t pixel = *ptr++;
-      if ((pixel & 0x00ffffff) != 0) {
-        max_y = std::max(max_y, static_cast<int32_t>(i));
-      }
-    }
-  }
-  return max_y;
-}
-}  // namespace
-
-// This test makes sure that given a tiny change in scale, a glyph will not do a
-// large jump in its drawn y position. This was noticed when performing
-// animations as a problematic artifact. We tried to come up with a more
-// holistic quantification of the problem but haven't yet been able to.
-TEST_P(DlGoldenTest, TextJumpingTest) {
-  SetWindowSize(impeller::ISize(1024, 200));
-  impeller::Scalar font_size = 300;
-  auto callback = [&](impeller::Scalar scale) -> sk_sp<DisplayList> {
-    DisplayListBuilder builder;
-    DlPaint paint;
-    paint.setColor(DlColor::ARGB(1, 0, 0, 0));
-    builder.DrawPaint(paint);
-    builder.Scale(scale, scale);
-    // If you move this code to a playgrounds test the RenderTextInCanvasSkia
-    // signature is a bit different there, it will look like this:
-    //
-    // RenderTextInCanvasSkia(GetContext(), builder,
-    //                    "the quick brown fox jumped over the lazy dog!.?",
-    //                    "Roboto-Regular.ttf",
-    //                    TextRenderOptions{
-    //                        .font_size = font_size,
-    //                        .position = SkPoint::Make(100, 300),
-    //                    });
-    // Note: The ahem font just has full blocks in it.
-    RenderTextInCanvasSkia(&builder, "h", "Roboto-Regular.ttf",
-                           DlPoint::MakeXY(10, 300),
-                           TextRenderOptions{
-                               .font_size = font_size,
-                           });
-    return builder.Build();
-  };
-
-  std::unique_ptr<impeller::testing::Screenshot> right =
-      MakeScreenshot(callback(0.445));
-  if (!right) {
-    GTEST_SKIP() << "making screenshots not supported.";
-  }
-  std::unique_ptr<impeller::testing::Screenshot> left =
-      MakeScreenshot(callback(0.444));
-
-  int32_t left_max_y = CalculateMaxY(left.get());
-  int32_t right_max_y = CalculateMaxY(right.get());
-  int32_t y_diff = std::abs(left_max_y - right_max_y);
-  EXPECT_TRUE(y_diff <= 1) << "y diff: " << y_diff;
 }
 
 }  // namespace testing
